ARM64: dts: qcom: xiaomi: Remove display ramdump memory region
This region is the same as the early splash framebuffer initialized by
XBL. msm_drm uses it for ramdump when available, but that's a debugging
feature we don't need. It also wastes memory after early boot because
msm_drm will keep it reserved rather than freeing the framebuffer memory
once it has taken control of the display hardware.

Remove the display ramdump region to save 35 MiB of memory

Change-Id: Icba33c282dfc4e263c2b300022afb36276bd940b
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: raghavt20 <raghavtandon2000@gmail.com>